requirements.txt
streamlit>=1.25.0
google-genai>=1.0.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
requests>=2.28.0
rich>=12.0.0

# CrewAI dependencies
crewai>=0.50.0
crewai-tools>=0.13.0
